Added UserService and modified UserDao a tiny bit for more desirable returns that UserService can use

+++ src/main/java/com/codecool/restauratio/dao/UserDao.java
@@ -19,7 +19,7 @@ public class UserDao {
         }
     }
 
-    public User getById(Integer userId) throws ConnectToDBFailed {
+    public User getById(long userId) throws ConnectToDBFailed {
         try {
             return em.find(User.class, userId);
         } catch (Exception e) {
